Desktop/trading-bot/run_professional_trading_bot.py
#!/usr/bin/env python3
"""Professional Trading Bot - Major overhaul based on expert analysis."""

import os
import sys
import asyncio
import logging
import signal
import traceback
import time
import numpy as np
import pandas as pd
from datetime import datetime, timedelta, timezone
from pathlib import Path
from typing import Dict, List, Optional, Any, Tuple
from dotenv import load_dotenv
from logging.handlers import RotatingFileHandler
import json

# Add src to path (will be removed for production packaging)
sys.path.append('src')

from src.config.settings import settings, app_settings
from src.data.market_data import MarketDataManager
from src.trading.broker_interface import BrokerManager
from src.analysis.technical import TechnicalAnalyzer, SignalDirection
from src.analysis.correlation import CorrelationAnalyzer
from src.scheduling.intelligent_scheduler import IntelligentTradingScheduler
from src.monitoring.metrics import MetricsCollector

# Professional logging with rotation
logs_dir = Path("logs")
logs_dir.mkdir(exist_ok=True)

# Rotating file handler (10MB max, 5 backups)
file_handler = RotatingFileHandler(
    logs_dir / "professional_trading_bot.log", 
    maxBytes=10*1024*1024, 
    backupCount=5, 
    encoding='utf-8'
)

logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(name)s - %(message)s',
    handlers=[file_handler, logging.StreamHandler()]
)

logger = logging.getLogger(__name__)

class ProfessionalTradingBot:
    """Professional-grade trading bot with expert-level optimizations."""
    
    def __init__(self):
        logger.info("🚀 INITIALIZING PROFESSIONAL TRADING BOT")
        
        # Core components
        self.data_manager = MarketDataManager()
        self.broker_manager = BrokerManager()
        self.technical_analyzer = TechnicalAnalyzer()
        self.correlation_analyzer = CorrelationAnalyzer(self.data_manager)
        self.scheduler = IntelligentTradingScheduler()
        self.metrics_collector = MetricsCollector()
        
        # Load parameters from settings (no hardcoding)
        self.trading_params = {
            'min_confidence': 0.78,
            'min_rr_ratio': 1.5,
            'profit_protection_percentage': getattr(settings.trading, 'profit_protection', 0.25),
            'max_volatility': 0.002,
            'minimum_profit_to_protect_pct': 0.005,  # 0.5% of equity
            **{f'atr_multiplier_{k}': v for k, v in app_settings.atr_multipliers.items()},
            **{f'volatility_threshold_{k}': v for k, v in app_settings.volatility_thresholds.items()}
        }
        
        # Trading state
        self.running = True
        self.current_mode = 'TRADING'
        self.position_trackers = {}
        self.account_info = {}
        
        # Performance tracking with timestamped heartbeat
        self.heartbeat_file = logs_dir / f"heartbeat_{datetime.now().strftime('%Y%m%d')}.json"
        self.scan_count = 0
        self.trades_executed = 0
        self.signals_analyzed = 0
        self.signals_rejected = 0
        self.positions_closed_profit = 0
        self.positions_closed_loss = 0
        
        # Currency pairs from settings
        self.currency_pairs = app_settings.CURRENCY_PAIRS.copy()
        
        # Pip values for proper position sizing
        self.pip_values = {
            'USDJPY': 0.01, 'EURJPY': 0.01, 'GBPJPY': 0.01, 'AUDJPY': 0.01,
            'CADJPY': 0.01, 'CHFJPY': 0.01, 'NZDJPY': 0.01
        }
        # Default for non-JPY pairs
        self.default_pip_value = 0.0001