Desktop/trading-bot/requirements.txt
# ===========================================
# Trading Bot Requirements
# Organized by environment and usage
# ===========================================

# ===========================================
# Trading Bot Requirements - PINNED VERSIONS
# Compatible with NumPy 2.0 and Gymnasium
# ===========================================

# Core Dependencies (Always Required)
python-dotenv==1.0.0
pandas==2.2.3
numpy==1.26.4  # Compatible with Gymnasium
MetaTrader5==5.0.45
pytz==2024.1
prometheus-client==0.20.0

# Technical Analysis
ta==0.11.0

# Async and Networking
aiohttp==3.10.5
requests==2.32.3

# Basic Sentiment Analysis
vaderSentiment==3.3.2

# Telegram Bot Control
python-telegram-bot==20.7

# ===========================================
# ML and Advanced Features (Optional)
# ===========================================

# Reinforcement Learning - PINNED FOR COMPATIBILITY
stable-baselines3==2.3.2  # Compatible with Gymnasium
gymnasium==0.29.1        # Modern gym replacement

# Machine Learning
scikit-learn==1.5.1
scipy==1.13.0

# Deep Learning - PINNED FOR COMPATIBILITY
torch==2.1.2             # Compatible with transformers
transformers==4.36.2     # Compatible with torch
accelerate==0.24.0       # Compatible with transformers

# ===========================================
# Development and Testing
# ===========================================

[dev]
pytest==8.3.2
pytest-asyncio==0.23.0
pytest-cov==5.0.0
black==24.8.0
ruff==0.6.0
mypy==1.11.0

# ===========================================
# Utilities (Optional)
# ===========================================

[utils]
beautifulsoup4==4.12.2
lxml==5.2.0
eventregistry==9.1
tweepy==4.14.0

# ===========================================
# Installation Commands
# ===========================================
# Core only: pip install -r requirements.txt -c core
# With ML: pip install -r requirements.txt -c core -c ml-optional
# Development: pip install -r requirements.txt -c core -c dev
# Full: pip install -r requirements.txt
